gardening package names and relocate some files
* LinkOverviewData & ArchiveListData are used at multiple sources, so I think it is placed data package :)
