+++ app/src/main/java/com/android/artic/ui/archive/ArchiveListAdapter.kt
@@ -5,10 +5,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.android.artic.R
-import com.android.artic.ui.archive.data.ArchiveListData
+import com.android.artic.data.archive.ArchiveListData
 
 class ArchiveListAdapter(val ctx: Context, val dataList: ArrayList<ArchiveListData>): RecyclerView.Adapter<ArchiveListAdapter.Holder>(){
