+++ app/src/main/java/com/android/artic/ui/search_result/SearchResultAdapter.kt
@@ -1,4 +1,4 @@
-package com.android.artic.ui
+package com.android.artic.ui.search_result
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
@@ -6,25 +6,26 @@ import androidx.fragment.app.FragmentStatePagerAdapter
 
 class SearchResultAdapter (fm:FragmentManager,private val num_fragmnet:Int) : FragmentStatePagerAdapter(fm) {
     companion object{
-        private var archiveResultFragment:ArchiveResultFragment?=null
-        private var linkResultFragment:LinkResultFragment?=null
+        private var archiveResultFragment: ArchiveResultFragment?=null
+        private var linkResultFragment: LinkResultFragment?=null
     }
 
-    fun getArchiveResultFragment () : ArchiveResultFragment{
-        if(archiveResultFragment==null)  archiveResultFragment=ArchiveResultFragment()
+    fun getArchiveResultFragment () : ArchiveResultFragment {
+        if(archiveResultFragment ==null)  archiveResultFragment =
+            ArchiveResultFragment()
         return archiveResultFragment!!
     }
 
-    fun getLinkReultFragment () : LinkResultFragment{
-        if(linkResultFragment==null)  linkResultFragment=LinkResultFragment()
+    fun getLinkReultFragment () : LinkResultFragment {
+        if(linkResultFragment ==null)  linkResultFragment = LinkResultFragment()
         return linkResultFragment!!
     }
 
 
     override fun getItem(position: Int): Fragment? {
         return when(position) {
-            0->ArchiveResultFragment()
-            1->LinkResultFragment()
+            0-> ArchiveResultFragment()
+            1-> LinkResultFragment()
             else->null
         }
