+++ app/src/main/res/layout/activity_link.xml
@@ -7,7 +7,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.LinkActivity"
+    tools:context=".ui.link.LinkActivity"
     android:background="#f2f2f2"
     >
