+++ app/src/main/res/layout/activity_search_result.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.SearchResultActivity">
+    tools:context=".ui.search_result.SearchResultActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
