+++ app/src/main/res/layout/fragment_archive_result.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.ArchiveResultFragment">
+    tools:context=".ui.search_result.ArchiveResultFragment">
 
 
     <androidx.recyclerview.widget.RecyclerView
